requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
bing-image-downloader==1.1.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.3.5
people-also-ask==1.1.0
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
setuptools==69.0.2
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
